Fix map loading bug

+++ CHAD/Assets/Scripts/Client/Server side/ServerSend.cs
@@ -264,7 +264,7 @@ public class ServerSend
     {
         using (Packet _packet = new Packet((int)ServerPackets.loadMap))
         {
-            _packet.Write(_mapType);
+            _packet.Write((int)_mapType);
             _packet.Write(_seed);
             SendTCPData(_toClient, _packet);
         }
